Fix bug that cleaned repos all the time
`doskey` doesn't allow you to use `&&` for multi-step commands; instead, it's `$T`. Having it as `&&` meant that `git clean -dfx` was getting run automatically every time the `AutoRun` registry key triggers, which it turns out is an awful lot. You'd expect it to only do it on `cmd.exe` startup but no, it does it when running any command-line programs too like npm.

Anyway, this seems to stop it automatically deleting all my uncommited files, which is nice.